auth-service/requirements.txt
psycopg2-binary==2.9.6
alembic==1.11.0
annotated-types==0.7.0
anyio==4.9.0
click==8.1.8
ecdsa==0.19.1
fastapi==0.115.12
h11==0.16.0
idna==3.10
passlib==1.7.4
pyasn1==0.4.8
pydantic==2.11.3
pydantic_core==2.33.1
python-jose==3.4.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
